refactor(marketplace): extract path-safe catalog lookups; add validatable
